# restaurant-service/web/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Restaurant API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/get-restaurants": {
      "get": {
        "summary": "Get list of restaurants",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "example": 1
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "example": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of restaurants",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer",
                            "example": 3
                          },
                          "title": {
                            "type": "string",
                            "example": "hotel test1"
                          },
                          "best_time": {
                            "type": "string",
                            "example": "https://www.bdstall.com/asset/Rangpur/9225_giant.jpg"
                          },
                          "picture_url": {
                            "type": "string",
                            "example": "test1 location"
                          },
                          "rating": {
                            "type": "integer",
                            "example": 0
                          },
                          "vote_count": {
                            "type": "integer",
                            "example": 0
                          },
                          "created_at": {
                            "type": "string",
                            "example": "2024-10-25T11:34:45.170623Z"
                          },
                          "updated_at": {
                            "type": "string",
                            "example": "2024-10-25T11:34:45.170624Z"
                          },
                          "is_active": {
                            "type": "boolean",
                            "example": false
                          }
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Success"
                    },
                    "status": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/get-restaurant/{id}": {
      "get": {
        "summary": "Get restaurant by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 3
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Details of a specific restaurant",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "example": "3"
                        },
                        "title": {
                          "type": "string",
                          "example": "hotel test1"
                        },
                        "content": {
                          "type": "string",
                          "example": "this is a nice test1"
                        },
                        "best_time": {
                          "type": "string",
                          "example": ""
                        },
                        "picture_url": {
                          "type": "string",
                          "example": "https://www.bdstall.com/asset/Rangpur/9225_giant.jpg"
                        },
                        "rating": {
                          "type": "integer",
                          "example": 0
                        },
                        "author": {
                          "type": "object",
                          "properties": {
                            "userid": {
                              "type": "integer",
                              "example": 1
                            },
                            "username": {
                              "type": "string",
                              "example": "mostofa"
                            }
                          }
                        },
                        "branches": {
                          "type": "string",
                          "nullable": true
                        },
                        "comments": {
                          "type": "string",
                          "nullable": true
                        },
                        "created_at": {
                          "type": "string",
                          "example": "0001-01-01T00:00:00Z"
                        },
                        "updated_at": {
                          "type": "string",
                          "example": "0001-01-01T00:00:00Z"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Success"
                    },
                    "status": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
  }